+++ JetFit/app/src/main/java/com/google/jetfit/presentation/screens/training/composables/ShowFullscreenIfChildrenAreFocused.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.platform.debugInspectorInfo
 
 @OptIn(ExperimentalFoundationApi::class)
-internal fun Modifier.bringIntoViewIfChildrenAreFocused(): Modifier = composed(
+internal fun Modifier.showFullscreenIfChildrenAreFocused(): Modifier = composed(
     inspectorInfo = debugInspectorInfo { name = "bringIntoViewIfChildrenAreFocused" },
     factory = {
         var myRect: Rect = Rect.Zero
@@ -23,12 +23,9 @@ internal fun Modifier.bringIntoViewIfChildrenAreFocused(): Modifier = composed(
             .bringIntoViewResponder(
                 remember {
                     object : BringIntoViewResponder {
-                        // return the current rectangle and ignoring the child rectangle received.
                         @ExperimentalFoundationApi
                         override fun calculateRectForParent(localRect: Rect): Rect = myRect
 
-                        // The container is not expected to be scrollable. Hence the child is
-                        // already in view with respect to the container.
                         @ExperimentalFoundationApi
                         override suspend fun bringChildIntoView(localRect: () -> Rect?) {
                         }
