+++ JetFit/app/src/main/java/com/google/jetfit/presentation/screens/training/composables/TrainingImageWithGradient.kt
@@ -1,4 +1,4 @@
-package com.google.jetfit.components
+package com.google.jetfit.presentation.screens.training.composables
 
 import android.content.res.Configuration
 import androidx.compose.foundation.Canvas
@@ -25,7 +25,7 @@ import coil.compose.AsyncImage
 
 @OptIn(ExperimentalTvMaterial3Api::class)
 @Composable
-fun ImageWithRadialGradient(
+fun TrainingImageWithGradient(
     modifier: Modifier = Modifier,
     imageUrl: String,
     contentDescription: String,
@@ -76,7 +76,7 @@ fun ImageWithRadialGradient(
 @Composable
 private fun ImageWithRadialGradientPreview() {
     Box(modifier = Modifier.fillMaxSize()) {
-        ImageWithRadialGradient(
+        TrainingImageWithGradient(
             Modifier.align(Alignment.TopEnd),
             "https://s3-alpha-sig.figma.com/img/5b76/28da/51c6b4c0076ea7b92c70d82dc1828425?Expires=1710720000&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4&Signature=M1NOL3tptpUBJ4qM0QbqmkCQBAKjWpNw3llpK70HUmcVyUPF9StrimFkFA32ziyk-X8GQz8VJHrT42ovbtj3ROiDwLBbLfpCbkuNaThYT5D0BAVkRZtSjkp~w3yDLQKdRSWp~1pn242mMj5ASFpYjL9udDM4JBHn9gjvzST7QGzvOHes9ZABFtimxVBC0Ot-eQDpDV7mbU9Pf5ROC2JTEd2LK-QG85N0Vv8cIFpUcPJGSFgR1tbHxMDv1GpKAx33eSGnH02~ow3R6sZm88wznn0AaPJoKwGGvU2ZJUVl6wbUD4JRt9gcs3q9FVFFEhSeoOpYbJSdqgWdzhPM-Lv-7Q__",
             "poster",
