+++ JetFit/app/src/main/java/com/google/jetfit/presentation/screens/training/series/SeriesScreen.kt
@@ -26,8 +26,8 @@ import androidx.tv.material3.MaterialTheme
 import androidx.tv.material3.OutlinedButton
 import androidx.tv.material3.Text
 import com.google.jetfit.R
-import com.google.jetfit.components.ImageWithRadialGradient
-import com.google.jetfit.presentation.screens.training.composables.MetricItem
+import com.google.jetfit.presentation.screens.training.composables.TrainingImageWithGradient
+import com.google.jetfit.presentation.screens.training.composables.TrainingInfoColumnItem
 import com.google.jetfit.presentation.theme.JetFitTheme
 
 @Composable
@@ -42,7 +42,7 @@ private fun SeriesScreenContent(
     state: SeriesUiState,
 ) {
     Box(modifier = Modifier.background(color = MaterialTheme.colorScheme.background)) {
-        ImageWithRadialGradient(
+        TrainingImageWithGradient(
             modifier = Modifier.align(Alignment.TopEnd),
             imageUrl = state.imageUrl,
             contentDescription = "Series poster"
@@ -76,10 +76,10 @@ private fun SeriesScreenContent(
                     modifier = Modifier.padding(top = 20.dp),
                     horizontalArrangement = Arrangement.spacedBy(40.dp)
                 ) {
-                    MetricItem(state.numberOfWeeks, "Weeks")
-                    MetricItem(state.numberOfClasses, "Classes")
-                    MetricItem(state.intensity, "Intensity")
-                    MetricItem(state.minutesPerDay, "Minutes per day")
+                    TrainingInfoColumnItem(state.numberOfWeeks, "Weeks")
+                    TrainingInfoColumnItem(state.numberOfClasses, "Classes")
+                    TrainingInfoColumnItem(state.intensity, "Intensity")
+                    TrainingInfoColumnItem(state.minutesPerDay, "Minutes per day")
                 }
                 Row(
                     modifier = Modifier.padding(top = 28.dp),
