+++ JetFit/app/src/main/java/com/google/jetfit/presentation/screens/training/workout/WorkoutScreen.kt
@@ -25,8 +25,8 @@ import androidx.tv.material3.Icon
 import androidx.tv.material3.MaterialTheme
 import androidx.tv.material3.Text
 import com.google.jetfit.R
-import com.google.jetfit.components.ImageWithRadialGradient
-import com.google.jetfit.presentation.screens.training.composables.MetricItem
+import com.google.jetfit.presentation.screens.training.composables.TrainingImageWithGradient
+import com.google.jetfit.presentation.screens.training.composables.TrainingInfoColumnItem
 import com.google.jetfit.presentation.theme.JetFitTheme
 
 @Composable
@@ -41,7 +41,7 @@ private fun WorkoutScreenContent(
     state: WorkoutUiState
 ) {
     Box(modifier = Modifier.background(color = MaterialTheme.colorScheme.background)) {
-        ImageWithRadialGradient(
+        TrainingImageWithGradient(
             Modifier.align(Alignment.TopEnd),
             state.imageUrl,
             "Workout image"
@@ -76,8 +76,8 @@ private fun WorkoutScreenContent(
                     modifier = Modifier.padding(top = 20.dp),
                     horizontalArrangement = Arrangement.spacedBy(24.dp)
                 ) {
-                    MetricItem(state.duration, "Duration")
-                    MetricItem(state.intensity, "Intensity")
+                    TrainingInfoColumnItem(state.duration, "Duration")
+                    TrainingInfoColumnItem(state.intensity, "Intensity")
                 }
                 Row(
                     modifier = Modifier.padding(top = 28.dp),
